requirements.txt
pandas
matplotlib
requests
bs4
lxml
nltk
numpy
plotly.express
wordcloud
h5py
geopy
streamlit_plotly_events
tweepy
altair

